requirements.txt
appdirs==1.4.3
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2017.4.17
chardet==2.3.0
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
Django==1.11.1
django-sendfile==0.3.11
feedparser==5.1.3
Flask==0.12.1
future==0.16.0
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
itsdangerous==0.24
Jinja2==2.9.6
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
oauthlib==1.0.3
onboard==1.2.0
packaging==16.8
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pygraphviz==1.3.1
PyJWT==1.3.0
pyparsing==2.2.0
pysendfile==2.0.1
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
python-telegram-bot==5.3.0
pytz==2017.2
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
transitions==0.5.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
Werkzeug==0.12.2
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
